Fix VC leave error / add various VC properties
This commit addresses an issue where an error would occur if a user left a voice channel and `.getUser()` returned `nil`.

This also fixes an issue where the wrong amount of members would be in `VoiceChannel.members` depending on the cache.

Make `VoiceChannel.userId` and `VoiceChannel.guildId` public. This releases the burden of depending on the `User` object being in cache for the state to properly update.